test_delete_link_not_found added.

+++ app/Http/Controllers/LinkController.php
@@ -63,7 +63,14 @@ class LinkController extends Controller
     public function destroy($id)
     {
         $link = Link::find($id);
+
+        if (!$link) {
+            return response()->json(['error' => 'Link not found'], 404);
+        }
+
         $link->delete();
+
         return response()->json([], 204);
     }
+
 }
